+++ android/src/main/java/com/amolg/flutterbarcodescanner/BarcodeCaptureActivity.java
@@ -31,11 +31,13 @@ import android.os.Bundle;
 
 import androidx.annotation.NonNull;
 
+import com.amolg.flutterbarcodescanner.constants.AppConstants;
 import com.google.android.material.snackbar.Snackbar;
 
 import androidx.core.app.ActivityCompat;
 import androidx.appcompat.app.AppCompatActivity;
 
+import android.util.DisplayMetrics;
 import android.util.Log;
 import android.view.GestureDetector;
 import android.view.MotionEvent;
@@ -64,7 +66,7 @@ import java.io.IOException;
  * size, and ID of each barcode.
  */
 public final class BarcodeCaptureActivity extends AppCompatActivity implements BarcodeGraphicTracker.BarcodeUpdateListener, View.OnClickListener {
-
+    private static final String TAG = BarcodeCaptureActivity.class.getSimpleName();
     // intent request code to handle updating play services if needed.
     private static final int RC_HANDLE_GMS = 9001;
 
@@ -87,6 +89,14 @@ public final class BarcodeCaptureActivity extends AppCompatActivity implements B
     private ImageView imgViewBarcodeCaptureUseFlash;
     private Button btnBarcodeCaptureCancel;
 
+    public static int SCAN_MODE = SCAN_MODE_ENUM.QR.ordinal();
+
+    public enum SCAN_MODE_ENUM {
+        QR,
+        BARCODE,
+        DEFAULT
+    }
+
     enum USE_FLASH {
         ON,
         OFF
@@ -102,6 +112,7 @@ public final class BarcodeCaptureActivity extends AppCompatActivity implements B
         super.onCreate(icicle);
         try {
             setContentView(R.layout.barcode_capture);
+
             String buttonText = "";
             try {
                 buttonText = (String) getIntent().getStringExtra("cancelButtonText");
@@ -133,6 +144,7 @@ public final class BarcodeCaptureActivity extends AppCompatActivity implements B
 
             gestureDetector = new GestureDetector(this, new CaptureGestureListener());
             scaleGestureDetector = new ScaleGestureDetector(this, new ScaleListener());
+
         } catch (Exception e) {
         }
     }
@@ -328,6 +340,7 @@ public final class BarcodeCaptureActivity extends AppCompatActivity implements B
                 mCameraSource = null;
             }
         }
+        System.gc();
     }
 
     /**
@@ -378,24 +391,46 @@ public final class BarcodeCaptureActivity extends AppCompatActivity implements B
         int i = v.getId();
         if (i == R.id.imgViewBarcodeCaptureUseFlash &&
                 getPackageManager().hasSystemFeature(PackageManager.FEATURE_CAMERA_FLASH)) {
-
             try {
                 if (flashStatus == USE_FLASH.OFF.ordinal()) {
                     flashStatus = USE_FLASH.ON.ordinal();
                     imgViewBarcodeCaptureUseFlash.setImageResource(R.drawable.ic_barcode_flash_on);
+                    turnOnOffFlashLight(true);
                 } else {
                     flashStatus = USE_FLASH.OFF.ordinal();
                     imgViewBarcodeCaptureUseFlash.setImageResource(R.drawable.ic_barcode_flash_off);
+                    turnOnOffFlashLight(false);
                 }
-
-                createCameraSource(true, (flashStatus == USE_FLASH.ON.ordinal()));
-                startCameraSource();
             } catch (Exception e) {
                 Toast.makeText(this, "Unable to turn on flash", Toast.LENGTH_SHORT).show();
                 Log.e("BarcodeCaptureActivity", "FlashOnFailure: " + e.getLocalizedMessage());
             }
         } else if (i == R.id.btnBarcodeCaptureCancel) {
+            Barcode barcode = new Barcode();
+            barcode.rawValue = "-1";
+            barcode.displayValue = "-1";
+            FlutterBarcodeScannerPlugin.onBarcodeScanReceiver(barcode);
             finish();
+        }
+    }
+
+    /**
+     * Turn on and off flash light based on flag
+     *
+     * @param isFlashToBeTurnOn
+     */
+    private void turnOnOffFlashLight(boolean isFlashToBeTurnOn) {
+        try {
+            if (getPackageManager().hasSystemFeature(PackageManager.FEATURE_CAMERA_FLASH)) {
+                String flashMode = "";
+                flashMode = isFlashToBeTurnOn ? Camera.Parameters.FLASH_MODE_TORCH : Camera.Parameters.FLASH_MODE_OFF;
+
+                mCameraSource.setFlashMode(flashMode);
+            } else {
+                Toast.makeText(getBaseContext(), "Unable to access flashlight as flashlight not available", Toast.LENGTH_SHORT).show();
+            }
+        } catch (Exception e) {
+            Toast.makeText(getBaseContext(), "Unable to access flashlight.", Toast.LENGTH_SHORT).show();
         }
     }