+++ lib/flutter_barcode_scanner.dart
@@ -2,6 +2,9 @@ import 'dart:async';
 
 import 'package:flutter/services.dart';
 
+/// Scan mode which is either QR code or BARCODE
+enum ScanMode { QR, BARCODE, DEFAULT }
+
 /// Flutter barcode scanner class that bridge the native classes to flutter project
 class FlutterBarcodeScanner {
   /// Create a method channel instance
@@ -17,17 +20,22 @@ class FlutterBarcodeScanner {
   /// lineColor is color of a line in scanning
   /// cancelButtonText is text of cancel button
   /// isShowFlashIcon is bool to show or hide flash icon
-  static Future<String> scanBarcode(
-      String lineColor, String cancelButtonText, bool isShowFlashIcon) async {
+  static Future<String> scanBarcode(String lineColor, String cancelButtonText,
+      bool isShowFlashIcon, ScanMode scanMode) async {
     if (null == cancelButtonText || cancelButtonText.isEmpty)
       cancelButtonText = "Cancel";
+
+    if (scanMode == null) {
+      scanMode = ScanMode.QR;
+    }
 
     /// create params to be pass to plugin
     Map params = <String, dynamic>{
       "lineColor": lineColor,
       "cancelButtonText": cancelButtonText,
       "isShowFlashIcon": isShowFlashIcon,
-      "isContinuousScan": false
+      "isContinuousScan": false,
+      "scanMode": scanMode.index
     };
 
     /// Get barcode scan result
@@ -41,17 +49,22 @@ class FlutterBarcodeScanner {
   /// This method allows continuous barcode scanning without closing camera.
   /// It will return stream of barcode strings.
   /// Parameters will be same as #scanBarcode
-  static Stream getBarcodeStreamReceiver(
-      String lineColor, String cancelButtonText, bool isShowFlashIcon) {
+  static Stream getBarcodeStreamReceiver(String lineColor,
+      String cancelButtonText, bool isShowFlashIcon, ScanMode scanMode) {
     if (null == cancelButtonText || cancelButtonText.isEmpty)
       cancelButtonText = "Cancel";
+
+    if (scanMode == null) {
+      scanMode = ScanMode.QR;
+    }
 
     /// create params to be pass to plugin
     Map params = <String, dynamic>{
       "lineColor": lineColor,
       "cancelButtonText": cancelButtonText,
       "isShowFlashIcon": isShowFlashIcon,
-      "isContinuousScan": true
+      "isContinuousScan": true,
+      "scanMode": scanMode.index
     };
 
     /// Invoke method to open camera